% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontpi.R
\name{decisiontpi}
\alias{decisiontpi}
\title{Dosing Decisions based on different tpi designs}
\usage{
decisiontpi(pt, e1 = 0.05, e2 = 0.05, x, n, eta, design = c("tpi",
  "mtpi", "mmtpi"), w, a1 = NULL, b1 = NULL, a2 = NULL, b2 = NULL)
}
\arguments{
\item{pt}{Target toxicity proportion to achieve in the sample (lower toxicity implies underdosing)}

\item{e1}{Allowable deviation towards the left of target toxicity \code{pt}, usually lies between 0 to 0.05}

\item{e2}{Allowable deviation towards the right of target toxicity \code{pt}, usually lies between 0 to 0.05}

\item{x}{Number of patients experiencing DLT (Dose Limiting Toxicity) 's in the cohort}

\item{n}{Total number of patients in the cohort}

\item{eta}{threshold value to check if the Dose is severely toxic}

\item{design}{Design parameter, tells us which design to use. Options are \code{"tpi", "mtpi", "mmtpi"}}

\item{w}{weight on the informative prior}

\item{a1, b1}{alpha and beta parameters for informative Beta Prior component}

\item{a2, b2}{alpha and Beta parameters for noninformative Beta Prior component}
}
\value{
\code{"E"} We should increase the current level of Dose.\cr
\code{"S"} We should stay at the current level of Dose and treat more patients.\cr
\code{"D"} We should decrease the current level of Dose.\cr
\code{"DU"} We should decrease the current level of Dose and we should not go beyond this Dose level for treating more patients\cr
}
\description{
\code{decisiontpi} gives whether to escalate, stay or de-escalate to a level of Dose given total number of patients and total number of people experiencing toxicities in a cohort
}
