% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upm.R
\name{UPM}
\alias{UPM}
\title{Calculation of Unit Probability Mass}
\usage{
UPM(w, a = 0, b = 1, a1 = 1, b1 = 1, a2 = 1, b2 = 1)
}
\arguments{
\item{w}{Weight on the first component of mixture distribution, i.e, the informative Prior}

\item{a, b}{Range Parameters}

\item{a1, b1}{alpha and beta parameters for informative prior component, must be input when w equals 1}

\item{a2, b2}{alpha and beta parameters for noninformative prior component, must be input when w equals 0}
}
\value{
Unit Probability Mass value or the UPM value
}
\description{
\code{UPM} calculates Unit Probability Mass for an interval (a, b) when the Underlying distribution is beta or mixture of two beta distributions.
}
\details{
UPM(a,b) = \eqn{(F(b) - F(a))/(b - a)}, defined for an interval (a,b), when X~F().
In this function, F() is assumed to be Cumulative Beta distribution function or mixture of two cumulative Beta distribution functions.
If F() consists of a single Beta distribution, and not a mixture, then we must either input \eqn{w = 1} and a1, b1 , or \eqn{w = 0} and a2,b2
}
\examples{
UPM(w = 1, a = 0.3, b = 0.4, a1 = 2, b1 = 5)
UPM(w = 0, a = 0.3, b = 0.4, a2 = 2, b2 = 5)
UPM(w = 0.3, a = 0.3, b = 0.4, a1 = 3, b1 = 6, a2 = 2, b2 = 5)
}
