% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upm.R
\name{upmplot}
\alias{upmplot}
\title{Graphical plot of Unit Probability MASS}
\usage{
upmplot(x, n, pt, e1 = 0.05, e2 = 0.05, design = c("mtpi", "mmtpi"),
  w, a1 = NULL, b1 = NULL, a2 = NULL, b2 = NULL)
}
\arguments{
\item{x}{a}

\item{n}{b}

\item{pt}{c}

\item{e1}{d}

\item{e2}{e}

\item{design}{f}

\item{w}{g}

\item{a1}{h}

\item{b1}{i}

\item{a2}{j}

\item{b2}{k}
}
\value{
A graph that includes Probability Distributions of the Dose Limiting Toxocity Rate.
}
\description{
\code{upmplot} Produces a graphical plot of Unit Probability Mass for a given set of parameters.
}
\examples{
require(ggplot2)
upmplot(x = 5, n = 7, pt = 0.3, design = "mmtpi", w = 0.1, a1 = 1, a2 = 1, b1 = 4, b2 = 6)
}
