% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_formulate.R
\name{weights_formulate}
\alias{weights_formulate}
\title{Posterior distribution parameters for mixture priors}
\usage{
weights_formulate(w = 1, x, n, a1 = 1, b1 = 1, a2 = 1, b2 = 1)
}
\arguments{
\item{w}{Weight on the informative prior (Distribution coming from a belief)}

\item{x}{Total Number of events}

\item{n}{Trial size}

\item{a1}{alpha parameter for informative distribution}

\item{b1}{beta parameter for informative distribution}

\item{a2}{alpha parameter for non informative distribution}

\item{b2}{beta parameter for non informative distribution}
}
\value{
\code{weights}  Weights on posterior distribution components. The first value refers to the weight for informative part of the posterior distribution.

\code{param_inform}  Parameters (alpha, beta) for the informative distribution

\code{param_noninform}  Parameters (alpha, beta) for the non informative distribution
}
\description{
\code{weights_formulate} calculates posterior distribution paramters for a Binomial likelihood and mixture Beta prior
}
\examples{
weights_formulate(w = 1, x = 1, n = 3, a1 = 1, b1= 1, a2 = 1, b2 = 1)
}
