% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontpi.R
\name{decisiontpi}
\alias{decisiontpi}
\title{Dosing Decisions based on different tpi designs}
\usage{
decisiontpi(pt, e1 = 0.05, e2 = 0.05, x, n, eta, design = c("tpi",
  "mtpi", "mmtpi"), w, a1 = NULL, b1 = NULL, a2 = NULL, b2 = NULL)
}
\arguments{
\item{pt}{Target toxicity proportion to achieve in current Dose Level (Less Toxicity means under- dosing, where as more toxicity means over - dosing)}

\item{e1}{Amount of variation that can be allowed to the left of the pt value to conclude that target toxicity has been achieved.
Default value is 0.05}

\item{e2}{Amount of variation that can be allowed to the right of the pt value to conclude that target toxicity has been achieved.
Default value is 0.05}

\item{x}{Total Number of events (In Dose Escalation Oncology Trials, this may be defined as
number of people who have experienced Dose Limiting Toxicities through administration of current Dose Level)}

\item{n}{Trial size (In Dose Escalation Oncology Trials, this may be defined as
total number of people who have been administered current Dose Level (missing responses will be excluded). Necessarily n will be greater than or equal to x}

\item{eta}{threshold value to check if the Dose is severely toxic}

\item{design}{Design parameter, tells us which design to use. Options are \code{"tpi", "mtpi", "mmtpi"}}

\item{w}{Weight on the first Beta distribution of the mixture Prior}

\item{a1}{alpha parameter ( > 0) for 1st Beta distribution, must be input properly when \eqn{w = 0 or 1}}

\item{b1}{beta parameter ( > 0) for 1st Beta distribution, must be input properly properly when \eqn{w = 0 or 1}}

\item{a2}{alpha parameter ( > 0) for 2nd Beta distribution, will not be used if \eqn{w = 0 or 1}}

\item{b2}{beta parameter ( > 0) for 2nd Beta distribution, will not be used if \eqn{w = 0 or 1}}
}
\value{
\code{"E"} We should increase the current level of Dose.\cr
\code{"S"} We should stay at the current level of Dose and treat more patients.\cr
\code{"D"} We should decrease the current level of Dose.\cr
\code{"DU"} We should decrease the current level of Dose and we should not go beyond this Dose level for treating more patients
}
\description{
\code{decisiontpi} gives whether to escalate, stay or de-escalate to a level of Dose given total number of patients and total number of people experiencing toxicities in a cohort
}
\details{
\code{decisiontpi} checks if the DLT rate within the sample at current Dose level is severely toxic. At first, it calculates the posterior distribution
of the DLT(or, Dose Limiting Toxicity) Rate based on the number of DLT 's in the sample. If the probability is too much (> eta), then the
function returns the value "DU", which implies that the Current Dose level is unacceptably toxic and can't ever be used for further administration.\cr
If the dose is not severely toxic, then it gives us the decision rule based on the design we provide.\cr
If we work with 'tpi' design, then the range of DLT rate, ie, [0,1], is broken up into three intervals, Under-Dosing [0, pt - e1), Target-Toxicity [pt - e1, pt + e2] and Over-Dosing (pt + e2, 1]\cr
Then, probability for these intervals are calculated and the interval for which the probability is high, leads to the Decision- making.\cr
For mTPI and mTPI-2 (coded as \code{"mmtpi"} in the package), the decision making after ensuring the non-severe toxicity of the current dose level
is based on the Unit Probability Mass among the Intervals.
}
\seealso{
\code{\link{UPM}} for definition on Unit Probability Mass, \code{\link{upmplot}} for the Decision Making Criterion based on UPM
}
