% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpitable.R
\name{tpitable}
\alias{tpitable}
\title{Title}
\usage{
tpitable(nmax, design, pt, e1 = 0.05, e2 = 0.05, eta, w, a1 = NULL,
  b1 = NULL, a2 = NULL, b2 = NULL)
}
\arguments{
\item{nmax}{A}

\item{design}{B}

\item{pt}{C}

\item{e1}{D}

\item{e2}{E}

\item{eta}{F}

\item{w}{G}

\item{a1}{H}

\item{b1}{T}

\item{a2}{G}

\item{b2}{G}
}
\value{
g
}
\description{
Title
}
