% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upm.R
\name{UPM}
\alias{UPM}
\title{Calculation of Unit Probability Mass}
\usage{
UPM(w, a = 0, b = 1, a1 = NULL, b1 = NULL, a2 = NULL,
  b2 = NULL)
}
\arguments{
\item{w, a1, b1, a2, b2}{Parameters for Prior Distribution. See \code{\link{weights_formulate}} for more information}

\item{a, b}{Range Parameters between which UPM is needed to be calculated.}
}
\value{
Unit Probability Mass value or the UPM value for the interval (a, b)
}
\description{
\code{UPM} calculates Unit Probability Mass for an interval (a, b) when the Underlying distribution is beta or mixture of two beta distributions.
}
\details{
Unit Probability MASS or UPM(a,b) = \eqn{(F(b) - F(a))/(b - a)}, defined for an interval (a,b), when X~F().
In this function, F() is assumed to be Cumulative Beta distribution function or mixture of two cumulative Beta distribution functions.

Hence, \eqn{F(x) =  w * pbeta(x, a1, b1) + (1 - w) * pbeta(x, a2, b2)}, pbeta is cumulative Beta distribution

If F() consists of a single Beta distribution, and not a mixture, then the convention here assumed is
to input \eqn{w = 1} and a1, b1 , or \eqn{w = 0} and a2,b2
}
\examples{
UPM(w = 1, a = 0.3, b = 0.4, a1 = 2, b1 = 5)
UPM(w = 0, a = 0.3, b = 0.4, a2 = 2, b2 = 5)
UPM(w = 0.3, a = 0.3, b = 0.4, a1 = 3, b1 = 6, a2 = 2, b2 = 5)
UPM(w = 1, a = 0.3, b = 0.4, a1 = 2, b1 = 5, a2 = 7, b2 = 8) #will give warning
}
\seealso{
\code{\link{weights_formulate}}, \code{\link[stats]{Beta}}
}
